Add checking to see if the fullsize image exists when unlinking images during reupload

+++ app/Services/CharacterManager.php
@@ -767,7 +767,7 @@ class CharacterManager extends Service
                 // Remove old versions so that images in various filetypes don't pile up
                 unlink($image->imagePath . '/' . $image->imageFileName);
                 unlink($image->imagePath . '/' . $image->fullsizeFileName);
-                unlink($image->imagePath . '/' . $image->thumbnailFileName);
+                if(file_exists( public_path($image->imageDirectory.'/'.$image->fullsizeFileName))) unlink($image->imagePath . '/' . $image->thumbnailFileName);
 
                 $image->extension = Config::get('lorekeeper.settings.masterlist_image_format');
                 $image->save();
